🛠 REFACTOR: Create handlers and log functions
